Rename SetUp to Bootstrap namespace
