+++ tests/Feature/Framework/BindingInterfacesWithInnerDependencies/FeatureTest.php
@@ -4,18 +4,17 @@ declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\BindingInterfacesWithInnerDependencies;
 
+use Gacela\Framework\Bootstrap\GacelaConfig;
 use Gacela\Framework\Gacela;
-use Gacela\Framework\Setup\SetupGacela;
 use PHPUnit\Framework\TestCase;
 
 final class FeatureTest extends TestCase
 {
     public function setUp(): void
     {
-        $setup = (new SetupGacela())
-            ->setExternalServices(['isWorking?' => 'yes!']);
-
-        Gacela::bootstrap(__DIR__, $setup);
+        Gacela::bootstrap(__DIR__, static function (GacelaConfig $config): void {
+            $config->addExternalService('isWorking?', 'yes!');
+        });
     }
 
     public function test_mapping_interfaces_from_config(): void
