+++ tests/Feature/Framework/UsingConfigTypePhp/gacela.php
@@ -2,12 +2,8 @@
 
 declare(strict_types=1);
 
-use Gacela\Framework\Config\GacelaConfigBuilder\ConfigBuilder;
-use Gacela\Framework\Setup\SetupGacela;
+use Gacela\Framework\Bootstrap\GacelaConfig;
 
-return (new SetupGacela())
-    ->setConfigFn(
-        static function (ConfigBuilder $configBuilder): void {
-            $configBuilder->add('config/*.php', 'config/local.php');
-        }
-    );
+return static function (GacelaConfig $config): void {
+    $config->addAppConfig('config/*.php', 'config/local.php');
+};
