+++ tests/Feature/Framework/UsingGacelaConfigFn/FeatureTest.php
@@ -4,14 +4,19 @@ declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\UsingGacelaConfigFn;
 
+use Gacela\Framework\Bootstrap\SetupGacela;
 use Gacela\Framework\Gacela;
+use GacelaTest\Fixtures\CustomClass;
 use PHPUnit\Framework\TestCase;
 
 final class FeatureTest extends TestCase
 {
     public function setUp(): void
     {
-        Gacela::bootstrap(__DIR__);
+        $setup = (new SetupGacela())
+            ->setExternalServices(['CustomClassFromExternalService' => CustomClass::class]);
+
+        Gacela::bootstrap(__DIR__, $setup);
     }
 
     public function test_config_php_files(): void
