+++ tests/Integration/Framework/Config/ConfigFactory/WithGacelaFile/gacela.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use Gacela\Framework\Setup\GacelaConfig;
+use Gacela\Framework\Bootstrap\GacelaConfig;
 use GacelaTest\Fixtures\CustomInterface;
 
 return static function (GacelaConfig $config): void {
