+++ tests/Unit/Framework/Config/GacelaFileConfig/Factory/GacelaConfigUsingGacelaPhpFileFactoryTest.php
@@ -4,16 +4,14 @@ declare(strict_types=1);
 
 namespace GacelaTest\Unit\Framework\Config\GacelaFileConfig\Factory;
 
+use Gacela\Framework\Bootstrap\GacelaConfig;
+use Gacela\Framework\Bootstrap\SetupGacela;
+use Gacela\Framework\Bootstrap\SetupGacelaInterface;
 use Gacela\Framework\Config\ConfigReader\PhpConfigReader;
 use Gacela\Framework\Config\FileIoInterface;
-use Gacela\Framework\Config\GacelaConfigBuilder\MappingInterfacesBuilder;
-use Gacela\Framework\Config\GacelaConfigBuilder\SuffixTypesBuilder;
 use Gacela\Framework\Config\GacelaFileConfig\Factory\GacelaConfigUsingGacelaPhpFileFactory;
 use Gacela\Framework\Config\GacelaFileConfig\GacelaConfigFile;
 use Gacela\Framework\Config\GacelaFileConfig\GacelaConfigItem;
-use Gacela\Framework\Setup\GacelaConfig;
-use Gacela\Framework\Setup\SetupGacela;
-use Gacela\Framework\Setup\SetupGacelaInterface;
 use GacelaTest\Fixtures\CustomClass;
 use GacelaTest\Fixtures\CustomInterface;
 use PHPUnit\Framework\TestCase;
@@ -23,7 +21,10 @@ final class GacelaConfigUsingGacelaPhpFileFactoryTest extends TestCase
     public function test_exception_when_the_class_does_not_implements_setup_gacela_interface(): void
     {
         $fileIo = $this->createStub(FileIoInterface::class);
-        $fileIo->method('include')->willReturn(new class() {});
+        $fileIo->method('include')->willReturn(
+            new class() {
+            }
+        );
 
         $factory = new GacelaConfigUsingGacelaPhpFileFactory(
             'gacelaPhpPath',
@@ -35,11 +36,27 @@ final class GacelaConfigUsingGacelaPhpFileFactoryTest extends TestCase
         $factory->createGacelaFileConfig();
     }
 
-    public function test_gacela_file_does_not_override_anything_then_use_defaults(): void
+    public function test_gacela_file_using_setup_class_does_not_override_anything_then_use_defaults(): void
     {
         $fileIo = $this->createStub(FileIoInterface::class);
         $fileIo->method('existsFile')->willReturn(true);
         $fileIo->method('include')->willReturn(new SetupGacela());
+
+        $factory = new GacelaConfigUsingGacelaPhpFileFactory(
+            'gacelaPhpPath',
+            $this->createStub(SetupGacelaInterface::class),
+            $fileIo
+        );
+
+        self::assertEquals(new GacelaConfigFile(), $factory->createGacelaFileConfig());
+    }
+
+    public function test_gacela_file_using_callable_does_not_override_anything_then_use_defaults(): void
+    {
+        $fileIo = $this->createStub(FileIoInterface::class);
+        $fileIo->method('existsFile')->willReturn(true);
+        $fileIo->method('include')->willReturn(static function (GacelaConfig $config): void {
+        });
 
         $factory = new GacelaConfigUsingGacelaPhpFileFactory(
             'gacelaPhpPath',
@@ -73,15 +90,12 @@ final class GacelaConfigUsingGacelaPhpFileFactoryTest extends TestCase
     {
         $fileIo = $this->createStub(FileIoInterface::class);
         $fileIo->method('include')->willReturn(
-            (new SetupGacela())
-                ->setExternalServices(['externalServiceKey' => 'externalServiceValue'])
-                ->setMappingInterfacesFn(
-                    static function (MappingInterfacesBuilder $mappingInterfacesBuilder, array $externalServices): void {
-                        self::assertSame('externalServiceValue', $externalServices['externalServiceKey']);
-                        $mappingInterfacesBuilder->bind(CustomInterface::class, new CustomClass());
-                        $mappingInterfacesBuilder->bind(CustomInterface::class, CustomClass::class);
-                    },
-                )
+            static function (GacelaConfig $config): void {
+                $config->addExternalService('externalServiceKey', 'externalServiceValue');
+                self::assertSame('externalServiceValue', $config->getExternalService('externalServiceKey'));
+                $config->mapInterface(CustomInterface::class, new CustomClass());
+                $config->mapInterface(CustomInterface::class, CustomClass::class);
+            }
         );
 
         $factory = new GacelaConfigUsingGacelaPhpFileFactory(
@@ -100,12 +114,9 @@ final class GacelaConfigUsingGacelaPhpFileFactoryTest extends TestCase
     {
         $fileIo = $this->createStub(FileIoInterface::class);
         $fileIo->method('include')->willReturn(
-            (new SetupGacela())
-                ->setSuffixTypesFn(
-                    static function (SuffixTypesBuilder $suffixTypesBuilder): void {
-                        $suffixTypesBuilder->addDependencyProvider('Binding');
-                    }
-                )
+            static function (GacelaConfig $config): void {
+                $config->addSuffixTypeDependencyProvider('Binding');
+            }
         );
 
         $factory = new GacelaConfigUsingGacelaPhpFileFactory(
